requirements.txt
# Build-time dependencies for local packaging and GitHub Actions.
# Source execution itself uses the Python standard library.
Pillow>=11.0,<12
PyInstaller>=6.19,<7
tzdata>=2024.1,<2026
